+++ CHANGELOG.md
@@ -15,8 +15,9 @@ The official release will have a defined and more stable API. If you are already
   * `with-git` removes its workspace after all child-steps are finished
   * Display build status in history (#3)
 * API Changes
-  * `lambdacd.internal.execution/execute-steps` is now deprecated. use the public `lambdacd.core/execute-steps` instead.
+  * `lambdacd.internal.execution/execute-steps` is now deprecated. use the public `lambdacd.core/execute-steps` instead. (#7)
     `lambdacd.core/execute-steps` takes keyword-arguments instead of argument lists for optional parameters.
+  * `lambdacd.internal.execution/execute-step` is now deprecated. use the public `lambdacd.core/execute-step` instead. (#7)
 
 ## 0.2.1